+++ src/features/simulation/index.ts
@@ -1,2 +1,3 @@
+export * from './hooks';
 export * from './sagas';
 export * from './slice';
