+++ src/features/simulation/sagas.test.ts
@@ -1,6 +1,4 @@
 import { GenericPermissionController } from '@metamask/permission-controller';
-import { DialogType } from '@metamask/rpc-methods';
-import { text } from '@metamask/snaps-ui';
 import { HandlerType } from '@metamask/snaps-utils';
 import { expectSaga } from 'redux-saga-test-plan';
 
@@ -12,45 +10,16 @@ import {
   permissionsSaga,
   rebootSaga,
   requestSaga,
-  showDialog,
 } from './sagas';
 import {
-  closeUserInterface,
-  getSnapName,
-  resolveUserInterface,
   sendRequest,
   setExecutionService,
   setManifest,
   setSourceCode,
-  showUserInterface,
 } from './slice';
 import { processSnapPermissions } from './snap-permissions';
 import { MockExecutionService } from './test/mockExecutionService';
 import { MOCK_MANIFEST } from './test/mockManifest';
-
-describe('showDialog', () => {
-  it('shows a dialog', async () => {
-    await expectSaga(showDialog, DEFAULT_SNAP_ID, DialogType.Alert, text('foo'))
-      .withState({
-        simulation: {
-          manifest: MOCK_MANIFEST,
-        },
-      })
-      .select(getSnapName)
-      .put(
-        showUserInterface({
-          snapId: DEFAULT_SNAP_ID,
-          snapName: '@metamask/example-snap',
-          type: DialogType.Alert,
-          node: text('foo'),
-        }),
-      )
-      .dispatch(resolveUserInterface('foo'))
-      .put(closeUserInterface())
-      .returns('foo')
-      .silentRun();
-  });
-});
 
 describe('initSaga', () => {
   it('initializes the execution environment', async () => {
